fix: Menu enum 으로 통합
