--- src/main/java/christmas/domain/Appetizer.java
@@ -1,23 +1,0 @@
-package christmas.domain;
-
-public enum Appetizer implements Menu {
-    MUSHROOM_SOUP("양송이수프", 6_000),
-    TAPAS("타파스", 5_500),
-    CAESAR_SALAD("시저샐러드", 8_000);
-
-    private final String menu;
-    private final int price;
-
-    Appetizer(String menu, int price) {
-        this.menu = menu;
-        this.price = price;
-    }
-
-    public String getMenu() {
-        return menu;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-}
