--- src/main/java/christmas/domain/Main.java
@@ -1,25 +1,0 @@
-package christmas.domain;
-
-public enum Main implements Menu {
-    T_BONE_STEAK("티본스테이크", 55_000),
-    BBQ_RIBS("바비큐립", 54_000),
-    SEAFOOD_PASTA("해산물파스타", 35_000),
-    CHRISTMAS_PASTA("크리스마스파스타", 25_000);
-
-
-    private final String menu;
-    private final int price;
-
-    Main(String menu, int price) {
-        this.menu = menu;
-        this.price = price;
-    }
-
-    public String getMenu() {
-        return menu;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-}
